% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{and_then}
\alias{and_then}
\title{Process a success or pass through a failure}
\usage{
and_then(x, fn, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Calls \code{fn()} on the wrapped value of \code{x} if \code{x} is a \code{Success}, or \code{Some},
otherwise returns \code{x} if it is a \code{Failure} or \code{Nothing}.

This function can be used for control flow based on Result values.
}
