% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{check_is_result}
\alias{check_is_result}
\title{Check if an object is a subclass of \code{Result}}
\usage{
check_is_result(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
Success(10) |> is_result()
Failure("OOps!") |> is_result()
Some(10) |> is_result()
Nothing() |> is_result()

}
\seealso{
\code{\link[=is_result]{is_result()}}
}
