% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{Failure}
\alias{Failure}
\title{A class for representing a failed or stopped computation}
\usage{
Failure(error = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
Failure("The computation failed")
Failure("Argument x must be numeric")
Failure(rlang::catch_cnd(stop("An error was thrown")))

}
\seealso{
\code{\link[=Success]{Success()}}, which wraps a value to provide the context of a successful computation.
}
