% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{inspect}
\alias{inspect}
\title{Apply a function to a \code{Result}'s value if it is a success}
\usage{
inspect(x, fn, ...)
}
\arguments{
\item{x}{A \code{Result} object.}

\item{fn}{A function to apply to the wrapped value.}

\item{...}{Additional optional values that may be used by specific methods.}
}
\value{
The input \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
