% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{is_failure_and}
\alias{is_failure_and}
\title{Check if an object is a subclass of \code{Failure} AND a predicate is true}
\usage{
is_failure_and(x, fn)
}
\arguments{
\item{x}{The object to test.}

\item{fn}{A single argument predicate function that accepts \code{x} as its argument.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
Success(10) |> is_failure_and(~ unwrap_fail(.x) == "OOps!")
Failure("OOps!") |> is_failure_and(~ unwrap_fail(.x) == "OOps!")
Some(10) |> is_failure_and(~ unwrap_fail(.x) == "OOps!")
Nothing() |> is_failure_and(~ unwrap_fail(.x) == "OOps!")

}
\seealso{
\code{\link[=is_success_and]{is_success_and()}}
}
