% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{is_success}
\alias{is_success}
\title{Check if an object is a subclass of \code{Success}}
\usage{
is_success(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
Check if an object is a subclass of \code{Success}
}
\examples{
Success(10) |> is_success()
Failure("OOps!") |> is_success()
Some(10) |> is_success()
Nothing() |> is_success()

}
\seealso{
\code{\link[=is_failure]{is_failure()}}
}
