% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{is_success_and}
\alias{is_success_and}
\title{Check if an object is a subclass of \code{Success} AND a predicate is true}
\usage{
is_success_and(x, fn)
}
\arguments{
\item{x}{The object to test.}

\item{fn}{A single argument predicate function that accepts \code{x} as its argument.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
Success(10) |> is_success_and(~ unwrap(.x) == 10)
Failure("OOps!") |> is_success_and(~ unwrap(.x) == 10)
Some(10) |> is_success_and(~ unwrap(.x) == 10)
Nothing() |> is_success_and(~ unwrap(.x) == 10)

}
\seealso{
\code{\link[=is_failure_and]{is_failure_and()}}
}
