% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Option.R
\name{Option}
\alias{Option}
\title{The abstract parent class of the \code{Some} and \code{Nothing} classes}
\usage{
Option()
}
\description{
As an \code{S7} abstract class. This can't actually be constructed, but can be used
for the purposes of dispatch.
}
