% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{p_cont_test}
\alias{p_cont_test}
\title{Perform significance testing on continuous data}
\usage{
p_cont_test(df, y, ..., subgrp = NULL, normal = FALSE)
}
\arguments{
\item{df}{data frame}

\item{y}{column which contains the corresponding groups, as a string or
symbol}

\item{...}{selection of column(s) with the data values, supports dplyr
selection rules}

\item{subgrp}{optional, additional columns used to subgroup the data prior to
testing as a character vector}

\item{normal}{logical, if TRUE will use t.test, otherwise wilcox.test}
}
\value{
tibble with the results of significance testing
}
\description{
Perform significance testing on continuous data
}
\keyword{internal}
