% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{Result}
\alias{Result}
\title{The abstract parent class of the \code{Success} and \code{Failure} classes}
\usage{
Result()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

As an \code{S7} abstract class. This can't actually be constructed, but can be used
for the purposes of dispatch.
}
