% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{Result}
\alias{Result}
\title{The abstract parent class of the \code{Success} and \code{Failure} classes.}
\usage{
Result()
}
\description{
The abstract parent class of the \code{Success} and \code{Failure} classes.
}
