% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultr-package.R
\docType{package}
\name{resultr-package}
\alias{resultr}
\alias{resultr-package}
\title{resultr: Result and Option S7 Classes, inspired by Rust}
\description{
S7 classes for 'Result' and 'Option' values that allow you to handle errors or missing values in a manner similar to that of recent systems programming languages like 'Rust'. Easily wrap existing functions into ones that return 'Result' or 'Option' values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://transgirlcodes.github.io/resultr/}
}

}
\author{
\strong{Maintainer}: Sabrina Ward \email{me@trans-girl.codes} [copyright holder]

}
\keyword{internal}
