% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{unwrap}
\alias{unwrap}
\title{Get the wrapped value from a \code{Result} or \code{Option} object}
\usage{
unwrap(x, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

If the provided object is a \code{Failure} (in the case of a \code{Result} input) or
\code{Nothing} (in the case of an \code{Option} input), then the function will 'panic'
and throw an error of class \code{unwrap_fail_panic}.

For this reason, use of this function is not generally advised unless you
specifically want to crash your R program or throw an error condition.

Instead you are encouraged to handle the "unhappy path" yourself using
\code{unwrap_or_default} or \code{unwrap_or_else}.
}
\examples{
Success(10) |> unwrap()
Success(1:10) |> unwrap()
Success(c(TRUE, TRUE, FALSE)) |> unwrap()
\dontrun{
Failure("Ooopsy") |> unwrap()
}

}
\seealso{
\code{\link[=unwrap_or_default]{unwrap_or_default()}} \code{\link[=unwrap_option]{unwrap_option()}} \code{\link[=unwrap_or_else]{unwrap_or_else()}}
}
