% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{unwrap_option}
\alias{unwrap_option}
\title{Get the wrapped value from a \code{Result}, wrapped in an \code{Option} object}
\usage{
unwrap_option(x, ...)
}
\arguments{
\item{x}{A \code{Result} object i.e. a \code{Success()} or \code{Failure()} value.}

\item{...}{Additional optional arguments that may be used by specific methods.}
}
\value{
\code{Some(x)} in the case of a \code{Success} where \code{x} is the unwrapped value of the \code{Success}, or \code{Nothing()} in the case of a \code{Failure()}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is useful for converting a \code{Result} to an \code{Option} i.e. to go from
the possibility of an answer or details of a failure, to the possibility of
an answer or nothing.
}
