% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{unwrap_or_default}
\alias{unwrap_or_default}
\title{Get the wrapped value from a \code{Result} or \code{Option} object}
\usage{
unwrap_or_default(x, default, ...)
}
\arguments{
\item{x}{The object to unwrap.}

\item{default}{The default value to return in the event that \code{x} is a \code{Failure}.}
}
\value{
The value wrapped by a \code{Success} or \code{default} in the case of a \code{Failure}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

If the provided object is a \code{Failure} (in the case of a \code{Result} input) or
\code{Nothing} (in the case of an \code{Option} input), then the function will return
the provided default value instead.
}
\examples{
Success(10) |> unwrap_or_default(200)
Failure("OMG I failed") |> unwrap_or_default(200)
Some(50) |> unwrap_or_default(200)
Nothing() |> unwrap_or_default(200)

}
\seealso{
\code{\link[=unwrap_or_else]{unwrap_or_else()}}
}
