% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{unwrap_or_else}
\alias{unwrap_or_else}
\title{Get the wrapped value from a \code{Result} or \code{Option} object}
\usage{
unwrap_or_else(x, fn, ...)
}
\arguments{
\item{x}{A \code{Result} or \code{Option} object.}

\item{fn}{A single argument function that accepts a \code{Failure} or a \code{Nothing}.}

\item{...}{Additional optional arguments that may be used by specific methods.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

If the provided object is a \code{Failure} (in the case of a \code{Result} input) or
\code{Nothing} (in the case of an \code{Option} input), then the provided function
\code{fn} shall be run instead.
}
\examples{
Success(10) |> unwrap_or_else(\(errval) { message("Oh no no I failed") })
Failure("Oh no I failed!") |> unwrap_or_else(\(errval) { message(errval) })

}
\seealso{
\code{\link[=unwrap_or_default]{unwrap_or_default()}}
}
