% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{is_failure}
\alias{is_failure}
\title{Check if an object is a subclass of \code{Failure}}
\usage{
is_failure(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
Success(10) |> is_failure()
Failure("OOps!") |> is_failure()
Some(10) |> is_failure()
Nothing() |> is_failure()

}
\seealso{
\code{\link[=is_success]{is_success()}}
}
