% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{result}
\alias{result}
\title{Wrap an existing function to return \code{Result} objects}
\usage{
result(fn, ...)
}
\arguments{
\item{fn}{A function to wrap.}

\item{...}{Additional optional arguments that may be used by individual methods.}
}
\description{
Wrap up an existing function so that it returns values wrapped in a \code{Success}
if it finished computing and returns a value, and returns a \code{Failure} if a
warning or error condition is intercepted during the functions execution.
}
\examples{
my_awesome_function <- function(do_the_throw = FALSE) {
    if(do_the_throw) {
        stop("Oh no!")
    } else {
        return(100)
    }
}
safe_awesome <- result(my_awesome_function)
safe_awesome()
safe_awesome(do_the_throw = TRUE)

}
