% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{unwrap_or_default}
\alias{unwrap_or_default}
\title{Get the wrapped value from a \code{Result} or \code{Option} object}
\usage{
unwrap_or_default(x, default, ...)
}
\description{
If the provided object is a \code{Failure} (in the case of a \code{Result} input) or
\code{Nothing} (in the case of an \code{Option} input), then the function will 'panic'
and throw an error of class \code{unwrap_fail_panic}.
}
\details{
For this reason, use of this function is not generally advised unless you
specifically want to crash your R program or throw an error condition.

Instead you are encouraged to handle the "unhappy path" yourself using
\code{unwrap_or_default} or \code{unwrap_or_else}.
}
\seealso{
\code{\link[=unwrap]{unwrap()}} \code{\link[=unwrap_or_else]{unwrap_or_else()}}
}
